backend/requirements.txt
asgiref==3.5.2
Django==3.2.13
Pillow==9.1.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
djangorestframework==3.12.2
gunicorn
